# apps/api/openapi/minerva.json
{
  "openapi": "3.1.1",
  "paths": {
...
    "/minerva/reviews/prompt/{promptId}": {
      "get": {
        "description": "Returns a single prompt of the caller's. Another user's prompt is not found.",
        "operationId": "DescribeReviewPrompt",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The prompt was found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DescribeReviewPromptResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Describes one of the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      },
      "put": {
        "description": "Rewords a prompt, changes or removes its placeholder, or archives it or brings it back. An archived prompt is no longer asked; reviews that answered it keep it.",
        "operationId": "UpdateReviewPrompt",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt to change",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Input for the UpdateReviewPrompt operation",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewPromptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The prompt with the changes applied.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateReviewPromptResponse"
                }
              }
            }
          },
          "304": {
            "description": "The request named nothing to change.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            }
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "summary": "Changes one of the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      },
      "delete": {
        "description": "Removes a prompt that has never been answered. A prompt with answers is refused; archive it instead, so the reviews that answered it keep it.",
        "operationId": "DeleteReviewPrompt",
        "parameters": [
          {
            "name": "promptId",
            "required": true,
            "in": "path",
            "description": "The ID of the prompt to delete",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The prompt was deleted."
          },
          "400": {
            "description": "The request presented was not valid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "No access token, or one that does not verify."
          },
          "404": {
            "description": "The entity with the specified identifiers was not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "The prompt has answers."
          }
        },
        "summary": "Deletes one of the signed-in user's review prompts",
        "tags": [
          "Review Prompts"
        ]
      }
    },
...
  },
  "info": {
    "title": "Minerva API",
    "description": "API documentation for Minerva",
    "version": "0.0.0",
    "contact": {
      "name": "Neil Fritz",
      "url": "https://ncfritz.net",
      "email": "ncfritz@ncfritz.net"
    }
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "statusCode": {
            "type": "number"
          }
        },
        "required": [
          "message",
          "statusCode"
        ]
      },
...
      "EmptyResponse": {
        "type": "object",
        "properties": {}
      },
...
      "ReviewKind": {
        "type": "string",
        "description": "Whether a review is of a day or of an ISO week",
        "enum": [
          "daily",
          "weekly"
        ]
      },
      "ReviewPromptSection": {
        "type": "string",
        "description": "The guided step a review prompt is asked in",
        "enum": [
          "reflect",
          "plan"
        ]
      },
      "ReviewPrompt": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "The unique ID of the prompt"
          },
          "kind": {
            "description": "The kind of review that asks the prompt",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewKind"
              }
            ]
          },
          "section": {
            "description": "The step the prompt is asked in",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPromptSection"
              }
            ]
          },
          "label": {
            "type": "string",
            "description": "The question, 1 to 120 characters"
          },
          "placeholder": {
            "type": "string",
            "description": "A hint shown in the empty answer box, at most 200 characters"
          },
          "position": {
            "type": "number",
            "description": "Where the prompt sits among its kind and section; lower comes first"
          },
          "archived": {
            "type": "boolean",
            "description": "Whether the prompt is archived: no longer asked, its answers kept"
          },
          "archivedTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the prompt was archived"
          },
          "createdTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the prompt was created"
          },
          "lastUpdatedTime": {
            "type": "string",
            "description": "An ISO-8601 formatted string indicating when the prompt was last changed"
          }
        },
        "required": [
          "id",
          "kind",
          "section",
          "label",
          "position",
          "archived",
          "createdTime"
        ]
      },
...
      "DescribeReviewPromptResponse": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The prompt.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "PartialReviewPrompt": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "description": "The question, 1 to 120 characters"
          },
          "placeholder": {
            "type": "string",
            "description": "A hint shown in the empty answer box, at most 200 characters"
          },
          "archived": {
            "type": "boolean",
            "description": "Whether the prompt is archived: no longer asked, its answers kept"
          }
        }
      },
      "UpdateReviewPromptRequest": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The changes to the prompt. A placeholder of null removes it; archived true archives it and false brings it back.",
            "allOf": [
              {
                "$ref": "#/components/schemas/PartialReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
      "UpdateReviewPromptResponse": {
        "type": "object",
        "properties": {
          "reviewPrompt": {
            "description": "The prompt with the changes applied.",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewPrompt"
              }
            ]
          }
        },
        "required": [
          "reviewPrompt"
        ]
      },
...
    }
  }
}